Ordering data previous to Pull Request.
