--- data-modelling/model_v2/list-DeviceType-to-DeviceType_target_encoded.json
@@ -1,4 +1,0 @@
-{
-    "desktop": 0.061458454725556465,
-    "mobile": 0.09888678080782683
-}
