--- data-modelling/model_v2/list-M3-to-M3_target_encoded.json
@@ -1,4 +1,0 @@
-{
-    "F": 0.03070350201443308,
-    "T": 0.017042003596565398
-}
