--- data-modelling/model_v2/list-M4-to-M4_target_encoded.json
@@ -1,5 +1,0 @@
-{
-    "M0": 0.03750353904690877,
-    "M1": 0.02668156192356655,
-    "M2": 0.11073320688076395
-}
