--- data-modelling/model_v2/list-M5-to-M5_target_encoded.json
@@ -1,4 +1,0 @@
-{
-    "F": 0.026427681632400935,
-    "T": 0.039161847882476494
-}
