--- data-modelling/model_v2/list-M6-to-M6_target_encoded.json
@@ -1,4 +1,0 @@
-{
-    "F": 0.023806808975413034,
-    "T": 0.017380113981870376
-}
