--- data-modelling/model_v2/list-P_emaildomain-to-P_emaildomain_1_target_encoded.json
@@ -1,62 +1,0 @@
-{
-    "aim.com": 0.16,
-    "anonymous.com": 0.02297388046513272,
-    "aol.com": 0.021285653469561516,
-    "att.net": 0.0047879616963064295,
-    "bellsouth.net": 0.02887844190732035,
-    "cableone.net": 0.02727272727272727,
-    "centurylink.net": 0.0,
-    "cfl.rr.com": 0.0,
-    "charter.net": 0.018154311649016642,
-    "comcast.net": 0.024832543702009477,
-    "cox.net": 0.017992424242424244,
-    "earthlink.net": 0.015665796344647518,
-    "embarqmail.com": 0.03015075376884422,
-    "frontier.com": 0.004807692307692308,
-    "frontiernet.net": 0.03597122302158273,
-    "gmail": 0.043943422188017445,
-    "gmail.com": 0.043943422188017445,
-    "gmx.de": 0.0,
-    "hotmail.co.uk": 0.05310207179602507,
-    "hotmail.com": 0.05310207179602507,
-    "hotmail.de": 0.05310207179602507,
-    "hotmail.es": 0.05310207179602507,
-    "hotmail.fr": 0.05310207179602507,
-    "icloud.com": 0.031163287086446105,
-    "juno.com": 0.0,
-    "live.com": 0.032127155901251266,
-    "live.com.mx": 0.032127155901251266,
-    "live.fr": 0.032127155901251266,
-    "mac.com": 0.026239067055393587,
-    "mail.com": 0.1880952380952381,
-    "me.com": 0.01240694789081886,
-    "msn.com": 0.015610066900286716,
-    "netzero.com": 0.0,
-    "netzero.net": 0.0,
-    "optonline.net": 0.008793969849246231,
-    "outlook.com": 0.09507722007722008,
-    "outlook.es": 0.09507722007722008,
-    "prodigy.net.mx": 0.006024096385542169,
-    "protonmail.com": 0.4603174603174603,
-    "ptd.net": 0.0,
-    "q.com": 0.0,
-    "roadrunner.com": 0.012552301255230125,
-    "rocketmail.com": 0.004048582995951417,
-    "sbcglobal.net": 0.00487156775907883,
-    "sc.rr.com": 0.008264462809917356,
-    "servicios-ta.com": 0.0,
-    "suddenlink.net": 0.0380952380952381,
-    "twc.com": 0.0,
-    "verizon.net": 0.006613131790269249,
-    "web.de": 0.0,
-    "windstream.net": 0.0,
-    "yahoo.co.jp": 0.0229777864127971,
-    "yahoo.co.uk": 0.0229777864127971,
-    "yahoo.com": 0.0229777864127971,
-    "yahoo.com.mx": 0.0229777864127971,
-    "yahoo.de": 0.0229777864127971,
-    "yahoo.es": 0.0229777864127971,
-    "yahoo.fr": 0.0229777864127971,
-    "ymail.com": 0.0221606648199446,
-    "NaN": NaN
-}
