--- data-modelling/model_v2/list-ProductCD-to-ProductCD_target_encoded.json
@@ -1,7 +1,0 @@
-{
-    "C": 0.11306976045655928,
-    "H": 0.04585122593156947,
-    "R": 0.03595686463866087,
-    "S": 0.061235216819973716,
-    "W": 0.02059387865352732
-}
