--- data-modelling/model_v2/list-R_emaildomain-to-R_emaildomain_1_target_encoded.json
@@ -1,63 +1,0 @@
-{
-    "aim.com": 0.03571428571428571,
-    "anonymous.com": 0.028095353928484553,
-    "aol.com": 0.03706050047513462,
-    "att.net": 0.0,
-    "bellsouth.net": 0.0,
-    "cableone.net": 0.0,
-    "centurylink.net": 0.0,
-    "cfl.rr.com": 0.0,
-    "charter.net": 0.046296296296296294,
-    "comcast.net": 0.013636363636363636,
-    "cox.net": 0.015306122448979591,
-    "earthlink.net": 0.09090909090909091,
-    "embarqmail.com": 0.0,
-    "frontier.com": 0.0,
-    "frontiernet.net": 0.0,
-    "gmail": 0.11345013656651788,
-    "gmail.com": 0.11345013656651788,
-    "gmx.de": 0.0,
-    "hotmail.co.uk": 0.07492950181281052,
-    "hotmail.com": 0.07492950181281052,
-    "hotmail.de": 0.07492950181281052,
-    "hotmail.es": 0.07492950181281052,
-    "hotmail.fr": 0.07492950181281052,
-    "icloud.com": 0.11915673693858846,
-    "juno.com": 0.0,
-    "live.com": 0.037539936102236424,
-    "live.com.mx": 0.037539936102236424,
-    "live.fr": 0.037539936102236424,
-    "mac.com": 0.0,
-    "mail.com": 0.3611111111111111,
-    "me.com": 0.002188183807439825,
-    "msn.com": 0.0014388489208633094,
-    "netzero.com": 0.0,
-    "netzero.net": 0.0,
-    "optonline.net": 0.012738853503184714,
-    "outlook.com": 0.15188172043010753,
-    "outlook.es": 0.15188172043010753,
-    "prodigy.net.mx": 0.006024096385542169,
-    "protonmail.com": 0.95,
-    "ptd.net": 0.0,
-    "q.com": 0.0,
-    "roadrunner.com": 0.0,
-    "rocketmail.com": 0.0,
-    "sbcglobal.net": 0.002109704641350211,
-    "sc.rr.com": 0.0,
-    "scranton.edu": 0.0,
-    "servicios-ta.com": 0.0,
-    "suddenlink.net": 0.0,
-    "twc.com": 0.0,
-    "verizon.net": 0.0,
-    "web.de": 0.0,
-    "windstream.net": 0.0,
-    "yahoo.co.jp": 0.04979004199160168,
-    "yahoo.co.uk": 0.04979004199160168,
-    "yahoo.com": 0.04979004199160168,
-    "yahoo.com.mx": 0.04979004199160168,
-    "yahoo.de": 0.04979004199160168,
-    "yahoo.es": 0.04979004199160168,
-    "yahoo.fr": 0.04979004199160168,
-    "ymail.com": 0.0449438202247191,
-    "NaN": NaN
-}
