--- data-modelling/model_v2/list-card3-to-card3_target_encoded.json
@@ -1,52 +1,0 @@
-{
-    "100.0": 0.04666666666666667,
-    "102.0": 0.02601156069364162,
-    "106.0": 0.012605042016806723,
-    "111.0": 0.0,
-    "117.0": 0.018207282913165267,
-    "118.0": 0.0,
-    "119.0": 0.09545454545454546,
-    "121.0": 0.0547945205479452,
-    "126.0": 0.0,
-    "127.0": 0.25,
-    "131.0": 0.0,
-    "133.0": 0.08823529411764706,
-    "134.0": 0.07913669064748201,
-    "135.0": 0.061855670103092786,
-    "137.0": 0.037037037037037035,
-    "138.0": 0.02631578947368421,
-    "141.0": 0.12,
-    "143.0": 0.03767660910518053,
-    "144.0": 0.06936416184971098,
-    "146.0": 0.009355509355509356,
-    "147.0": 0.02564102564102564,
-    "148.0": 0.0,
-    "149.0": 0.011111111111111112,
-    "150.0": 0.024859537578191603,
-    "159.0": 0.05714285714285714,
-    "162.0": 0.012396694214876033,
-    "166.0": 0.02857142857142857,
-    "168.0": 0.2,
-    "169.0": 0.0,
-    "171.0": 0.0,
-    "177.0": 0.06666666666666667,
-    "180.0": 0.125,
-    "182.0": 0.19047619047619047,
-    "185.0": 0.12597313908490781,
-    "188.0": 0.008064516129032258,
-    "190.0": 0.0,
-    "195.0": 0.01098901098901099,
-    "197.0": 0.013333333333333334,
-    "198.0": 0.21052631578947367,
-    "203.0": 0.0049504950495049506,
-    "204.0": 0.0,
-    "210.0": 0.011235955056179775,
-    "212.0": 0.0,
-    "213.0": 0.005555555555555556,
-    "214.0": 0.09523809523809523,
-    "217.0": 0.0,
-    "222.0": 0.0,
-    "223.0": 0.42424242424242425,
-    "227.0": 0.0,
-    "229.0": 0.010101010101010102
-}
