--- data-modelling/model_v2/list-card6-to-card6_target_encoded.json
@@ -1,4 +1,0 @@
-{
-    "credit": 0.06681659517565948,
-    "debit": 0.023997601463596703
-}
