--- data-modelling/model_v2/list-id_17-to-id_17_target_encoded.json
@@ -1,61 +1,0 @@
-{
-    "100.0": 0.1038961038961039,
-    "101.0": 0.0,
-    "102.0": 0.0823327615780446,
-    "106.0": 0.17857142857142858,
-    "107.0": 0.0,
-    "110.0": 1.0,
-    "111.0": 0.4,
-    "114.0": 0.0,
-    "118.0": 0.0,
-    "119.0": 0.0,
-    "121.0": 0.018433179723502304,
-    "127.0": 0.0,
-    "128.0": 0.0,
-    "131.0": 0.0,
-    "133.0": 0.023809523809523808,
-    "134.0": 0.0,
-    "135.0": 0.0,
-    "137.0": 0.0,
-    "138.0": 0.0,
-    "140.0": 0.0,
-    "141.0": 0.0,
-    "142.0": 0.0,
-    "144.0": 0.0,
-    "146.0": 0.0,
-    "147.0": 0.0,
-    "148.0": 0.04477611940298507,
-    "149.0": 0.0,
-    "150.0": 0.0,
-    "152.0": 0.0,
-    "155.0": 0.0,
-    "156.0": 0.0,
-    "159.0": 0.01384083044982699,
-    "162.0": 0.09090909090909091,
-    "163.0": 0.0,
-    "164.0": 0.0,
-    "166.0": 0.04181001897650308,
-    "168.0": 0.0,
-    "177.0": 0.0,
-    "183.0": 0.0,
-    "184.0": 0.0,
-    "185.0": 0.0,
-    "189.0": 0.0,
-    "190.0": 1.0,
-    "191.0": 0.009174311926605505,
-    "192.0": 0.0,
-    "195.0": 0.04878048780487805,
-    "197.0": 1.0,
-    "199.0": 0.42857142857142855,
-    "200.0": 0.0,
-    "203.0": 0.0,
-    "205.0": 0.047619047619047616,
-    "208.0": 0.0,
-    "210.0": 0.0,
-    "212.0": 0.0,
-    "216.0": 0.0,
-    "217.0": 0.0,
-    "218.0": 0.0,
-    "225.0": 0.12756191044502077,
-    "228.0": 0.0
-}
