--- data-modelling/model_v2/list-id_30-to-id_30_target_encoded.json
@@ -1,70 +1,0 @@
-{
-    "Android": 0.11847133757961784,
-    "Android 4.4.2": 0.1568627450980392,
-    "Android 5.0.2": 0.0,
-    "Android 5.1.1": 0.31493506493506496,
-    "Android 6.0": 0.05555555555555555,
-    "Android 6.0.1": 0.08571428571428572,
-    "Android 7.0": 0.05355064027939464,
-    "Android 7.1.1": 0.08856088560885608,
-    "Android 7.1.2": 0.0,
-    "Android 8.0.0": 0.0379746835443038,
-    "Android 8.1.0": 0.03333333333333333,
-    "Linux": 0.08381171067738231,
-    "Mac": 0.0,
-    "Mac OS X 10.10": 0.04310344827586207,
-    "Mac OS X 10.11": 0.022222222222222223,
-    "Mac OS X 10.12": 0.01639344262295082,
-    "Mac OS X 10.13": 0.0,
-    "Mac OS X 10.6": 0.0,
-    "Mac OS X 10.9": 0.0,
-    "Mac OS X 10_10_5": 0.017409470752089137,
-    "Mac OS X 10_11_3": 0.0,
-    "Mac OS X 10_11_4": 0.0625,
-    "Mac OS X 10_11_5": 0.0,
-    "Mac OS X 10_11_6": 0.010304219823356232,
-    "Mac OS X 10_12": 0.0,
-    "Mac OS X 10_12_1": 0.1282051282051282,
-    "Mac OS X 10_12_2": 0.0,
-    "Mac OS X 10_12_3": 0.0,
-    "Mac OS X 10_12_4": 0.0,
-    "Mac OS X 10_12_5": 0.01507537688442211,
-    "Mac OS X 10_12_6": 0.014831460674157304,
-    "Mac OS X 10_13_1": 0.01678445229681979,
-    "Mac OS X 10_13_2": 0.016260162601626018,
-    "Mac OS X 10_13_3": 0.05013927576601671,
-    "Mac OS X 10_13_4": 0.03571428571428571,
-    "Mac OS X 10_13_5": 0.0,
-    "Mac OS X 10_7_5": 0.0,
-    "Mac OS X 10_8_5": 0.0,
-    "Mac OS X 10_9_5": 0.045816733067729085,
-    "Windows 10": 0.03105326876513317,
-    "Windows 7": 0.0344115004492363,
-    "Windows 8": 0.14388489208633093,
-    "Windows 8.1": 0.0682095006090134,
-    "Windows Vista": 0.06363636363636363,
-    "Windows XP": 0.008620689655172414,
-    "iOS": 0.023715415019762844,
-    "iOS 10.0.2": 0.024096385542168676,
-    "iOS 10.1.1": 0.0,
-    "iOS 10.2.0": 0.05185185185185185,
-    "iOS 10.2.1": 0.08823529411764706,
-    "iOS 10.3.1": 0.03365384615384615,
-    "iOS 10.3.2": 0.047979797979797977,
-    "iOS 10.3.3": 0.06064880112834979,
-    "iOS 11.0.0": 0.09219858156028368,
-    "iOS 11.0.1": 0.20512820512820512,
-    "iOS 11.0.2": 0.05333333333333334,
-    "iOS 11.0.3": 0.03365384615384615,
-    "iOS 11.1.0": 0.1867704280155642,
-    "iOS 11.1.1": 0.07100591715976332,
-    "iOS 11.1.2": 0.03470893686799672,
-    "iOS 11.2.0": 0.08082026537997587,
-    "iOS 11.2.1": 0.04894722450095707,
-    "iOS 11.2.2": 0.12028301886792453,
-    "iOS 11.2.5": 0.11808809746954077,
-    "iOS 11.2.6": 0.059205190592051905,
-    "iOS 11.3.0": 0.078125,
-    "iOS 9.3.5": 0.015345268542199489,
-    "other": 0.3333333333333333
-}
