--- data-modelling/model_v2/list-id_31-to-id_31_target_encoded.json
@@ -1,87 +1,0 @@
-{
-    "Generic/Android": 0.06382978723404255,
-    "Generic/Android 7.0": 0.15151515151515152,
-    "Lanix/Ilium": 1.0,
-    "Microsoft/Windows": 0.0,
-    "Samsung/SM-G531H": 0.023255813953488372,
-    "Samsung/SM-G532M": 0.1523809523809524,
-    "android browser 4.0": 0.08974358974358974,
-    "android webview 4.0": 0.2727272727272727,
-    "chrome": 0.08831908831908832,
-    "chrome 43.0 for android": 0.23809523809523808,
-    "chrome 46.0 for android": 0.3018867924528302,
-    "chrome 49.0": 0.046801872074883,
-    "chrome 49.0 for android": 0.5833333333333334,
-    "chrome 50.0 for android": 0.2727272727272727,
-    "chrome 51.0": 0.0,
-    "chrome 51.0 for android": 0.08571428571428572,
-    "chrome 52.0 for android": 0.11904761904761904,
-    "chrome 53.0 for android": 0.0,
-    "chrome 54.0 for android": 0.12727272727272726,
-    "chrome 55.0": 0.19230769230769232,
-    "chrome 55.0 for android": 0.04285714285714286,
-    "chrome 56.0": 0.12643678160919541,
-    "chrome 56.0 for android": 0.16759776536312848,
-    "chrome 57.0": 0.0,
-    "chrome 57.0 for android": 0.05333333333333334,
-    "chrome 58.0": 0.25925925925925924,
-    "chrome 58.0 for android": 0.20952380952380953,
-    "chrome 59.0": 0.031496062992125984,
-    "chrome 59.0 for android": 0.06467661691542288,
-    "chrome 60.0": 0.08797653958944282,
-    "chrome 60.0 for android": 0.07023411371237458,
-    "chrome 61.0": 0.05492730210016155,
-    "chrome 61.0 for android": 0.1377952755905512,
-    "chrome 62.0": 0.04132810311011488,
-    "chrome 62.0 for android": 0.07763671875,
-    "chrome 63.0": 0.06796649732253192,
-    "chrome 63.0 for android": 0.1107773851590106,
-    "chrome 63.0 for ios": 0.11374407582938388,
-    "chrome 64.0": 0.0921293471629042,
-    "chrome 64.0 for android": 0.16365824308062576,
-    "chrome 64.0 for ios": 0.0,
-    "chrome 65.0": 0.11249628970020778,
-    "chrome 65.0 for android": 0.1619411123227917,
-    "chrome 65.0 for ios": 0.12195121951219512,
-    "chrome generic": 0.19600413365483982,
-    "chrome generic for android": 0.14388489208633093,
-    "comodo": 0.0,
-    "edge": 0.25,
-    "edge 13.0": 0.034482758620689655,
-    "edge 14.0": 0.06702412868632708,
-    "edge 15.0": 0.027433050293925537,
-    "edge 16.0": 0.02514427040395713,
-    "firefox": 0.2,
-    "firefox 47.0": 0.0967741935483871,
-    "firefox 48.0": 0.012121212121212121,
-    "firefox 52.0": 0.026785714285714284,
-    "firefox 55.0": 0.07142857142857142,
-    "firefox 56.0": 0.016666666666666666,
-    "firefox 57.0": 0.05356055995130858,
-    "firefox 58.0": 0.08913308913308914,
-    "firefox 59.0": 0.06553398058252427,
-    "firefox generic": 0.14285714285714285,
-    "ie": 0.06976744186046512,
-    "ie 11.0 for desktop": 0.018217821782178217,
-    "ie 11.0 for tablet": 0.0018726591760299626,
-    "mobile": 0.3333333333333333,
-    "mobile safari 10.0": 0.04294478527607362,
-    "mobile safari 11.0": 0.0584821833348445,
-    "mobile safari 8.0": 0.02040816326530612,
-    "mobile safari 9.0": 0.022,
-    "mobile safari generic": 0.1015728843597696,
-    "mobile safari uiwebview": 0.047058823529411764,
-    "opera": 0.5217391304347826,
-    "opera generic": 0.7777777777777778,
-    "other": 0.1890909090909091,
-    "safari": 0.0,
-    "safari generic": 0.022359084214754316,
-    "samsung": 0.0,
-    "samsung browser 4.0": 0.047619047619047616,
-    "samsung browser 4.2": 0.3333333333333333,
-    "samsung browser 5.2": 0.2857142857142857,
-    "samsung browser 5.4": 0.325,
-    "samsung browser 6.2": 0.07109004739336493,
-    "samsung browser 6.4": 0.0472972972972973,
-    "samsung browser 7.0": 0.0
-}
