--- data-modelling/model_v2/model_xgbv3_joblib_describe_X.csv
@@ -1,9 +1,0 @@
-,V264,V156,D3,V44,R_emaildomain_1_target_encoded,card6_target_encoded,V283,V62,V30,card3_target_encoded,V86,M4_target_encoded,C13,C5,M5_target_encoded,V198,DeviceType_target_encoded,V67,id_31_target_encoded,V54,V47,D2,id_17_target_encoded,V13,ProductCD_target_encoded,M3_target_encoded,V20,id_30_target_encoded,V171,id_01,V82,P_emaildomain_1_target_encoded,M6_target_encoded,TransactionAmt,V307,card2_freq_encoded,V312,D8,id_09,V162
-count,105261.0,67246.0,236769.0,312634.0,108240.0,442032.0,442631.0,378306.0,379313.0,441615.0,370570.0,229171.0,442905.0,442905.0,176164.0,111436.0,112395.0,378306.0,111699.0,378306.0,312634.0,224113.0,111256.0,379313.0,442905.0,213112.0,379313.0,63658.0,111436.0,115052.0,370570.0,374034.0,308059.0,442905.0,442893.0,442905.0,442893.0,58746.0,58746.0,67246.0
-mean,147.46400451660156,0.7820093631744385,28.65277549003459,1.0887491703033447,0.07858462675535845,0.03515356354291092,1.003933310508728,0.869735062122345,0.40410953760147095,0.03512561846857555,1.0662951469421387,0.05034232079975215,32.695331956062816,5.415628633679908,0.03214050543811448,0.9677303433418274,0.07602651363494817,1.0001426935195923,0.07549754250261863,0.5947777628898621,1.0405042171478271,166.29540455038307,0.07551952254260443,0.5772673487663269,0.03513846084374754,0.0199378730432824,0.8511229753494263,0.043325269408401146,1.5381295680999756,-9.891527309390536,0.8446636199951172,0.03609297550490062,0.020856394391983352,134.2162646278547,237.4404296875,0.03914134293593495,36.36254119873047,149.95689256268827,0.10521567425867293,6.6463541984558105
-std,2025.41015625,0.584498405456543,62.31665051157136,0.7045078873634338,0.04290311854725546,0.01879447736228636,1.5619667768478394,0.4735829532146454,0.5528890490531921,0.030751292032266234,0.4398348927497864,0.030465135501540366,136.23776957524694,25.200565939973874,0.00633339718955654,0.48172932863235474,0.018249183047877912,0.23880544304847717,0.05010211209200841,0.5347328782081604,0.24070821702480316,172.16707744781837,0.0437361447033013,0.5328152179718018,0.030058321281028744,0.005583547465682781,0.4629403054714203,0.03359810136074638,1.99747896194458,14.223173689948403,0.42279526591300964,0.016176746371860513,0.0032025773545756447,227.66548850867832,1564.725341796875,0.035737842840922115,171.59320068359375,236.02230502454512,0.9453233610775814,72.68865966796875
-min,0.0,0.0,0.0,0.0,0.0,0.023997601463596703,0.0,0.0,0.0,0.0,0.0,0.02668156192356655,0.0,0.0,0.026427681632400935,0.0,0.061458454725556465,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.02059387865352732,0.017042003596565398,0.0,0.0,0.0,-100.0,0.0,0.0,0.017380113981870376,0.251,0.0,3.0480577098926406e-05,0.0,0.0,-36.0,0.0
-25%,0.0,0.0,1.0,1.0,0.0449438202247191,0.023997601463596703,0.0,1.0,0.0,0.024859537578191603,1.0,0.03750353904690877,1.0,0.0,0.026427681632400935,1.0,0.061458454725556465,1.0,0.04132810311011488,0.0,1.0,24.0,0.04181001897650308,0.0,0.02059387865352732,0.017042003596565398,1.0,0.03105326876513317,1.0,-10.0,1.0,0.0229777864127971,0.017380113981870376,43.25,0.0,0.005652453686456464,0.0,0.9583330154418944,0.0,0.0
-50%,0.0,1.0,8.0,1.0,0.07492950181281052,0.023997601463596703,1.0,1.0,0.0,0.024859537578191603,1.0,0.03750353904690877,2.0,0.0,0.026427681632400935,1.0,0.061458454725556465,1.0,0.06796649732253192,1.0,1.0,97.0,0.04181001897650308,1.0,0.02059387865352732,0.017042003596565398,1.0,0.0344115004492363,1.0,-5.0,1.0,0.043943422188017445,0.023806808975413034,68.95,0.0,0.025722220340705117,0.0,41.20833206176758,0.0,0.0
-75%,29.732999801635742,1.0,27.0,1.0,0.11345013656651788,0.06681659517565948,1.0,1.0,1.0,0.024859537578191603,1.0,0.03750353904690877,12.0,1.0,0.039161847882476494,1.0,0.09888678080782683,1.0,0.1015728843597696,1.0,1.0,275.0,0.12756191044502077,1.0,0.03595686463866087,0.017042003596565398,1.0,0.04894722450095707,1.0,-5.0,1.0,0.043943422188017445,0.023806808975413034,125.0,139.9499969482422,0.07652487553764352,0.0,192.23958206176758,0.0,0.0
-max,153600.0,16.0,729.0,48.0,0.95,0.06681659517565948,68.0,7.0,9.0,0.42424242424242425,30.0,0.11073320688076395,2918.0,349.0,0.039161847882476494,14.0,0.09888678080782683,8.0,1.0,6.0,12.0,640.0,1.0,6.0,0.11306976045655928,0.03070350201443308,15.0,0.3333333333333333,61.0,0.0,7.0,0.4603174603174603,0.023806808975413034,6450.97,108800.0,0.09921597182239984,55125.0,1707.7916259765625,25.0,3300.0
