--- data-modelling/model_v2/model_xgbv3_joblib_describe_y.csv
@@ -1,9 +1,0 @@
-,isFraud
-count,442905.0
-mean,0.035138460843747533
-std,0.18412991598877712
-min,0.0
-25%,0.0
-50%,0.0
-75%,0.0
-max,1.0
