--- data-modelling/model_v2/model_xgbv3_joblib_info_X.txt
@@ -1,47 +1,0 @@
-<class 'pandas.core.frame.DataFrame'>
-Index: 442905 entries, 2987000 to 3429904
-Data columns (total 40 columns):
- #   Column                          Non-Null Count   Dtype  
----  ------                          --------------   -----  
- 0   V264                            105261 non-null  float32
- 1   V156                            67246 non-null   float32
- 2   D3                              236769 non-null  float64
- 3   V44                             312634 non-null  float32
- 4   R_emaildomain_1_target_encoded  108240 non-null  float64
- 5   card6_target_encoded            442032 non-null  float64
- 6   V283                            442631 non-null  float32
- 7   V62                             378306 non-null  float32
- 8   V30                             379313 non-null  float32
- 9   card3_target_encoded            441615 non-null  float64
- 10  V86                             370570 non-null  float32
- 11  M4_target_encoded               229171 non-null  float64
- 12  C13                             442905 non-null  float64
- 13  C5                              442905 non-null  float64
- 14  M5_target_encoded               176164 non-null  float64
- 15  V198                            111436 non-null  float32
- 16  DeviceType_target_encoded       112395 non-null  float64
- 17  V67                             378306 non-null  float32
- 18  id_31_target_encoded            111699 non-null  float64
- 19  V54                             378306 non-null  float32
- 20  V47                             312634 non-null  float32
- 21  D2                              224113 non-null  float64
- 22  id_17_target_encoded            111256 non-null  float64
- 23  V13                             379313 non-null  float32
- 24  ProductCD_target_encoded        442905 non-null  float64
- 25  M3_target_encoded               213112 non-null  float64
- 26  V20                             379313 non-null  float32
- 27  id_30_target_encoded            63658 non-null   float64
- 28  V171                            111436 non-null  float32
- 29  id_01                           115052 non-null  float64
- 30  V82                             370570 non-null  float32
- 31  P_emaildomain_1_target_encoded  374034 non-null  float64
- 32  M6_target_encoded               308059 non-null  float64
- 33  TransactionAmt                  442905 non-null  float64
- 34  V307                            442893 non-null  float32
- 35  card2_freq_encoded              442905 non-null  float64
- 36  V312                            442893 non-null  float32
- 37  D8                              58746 non-null   float64
- 38  id_09                           58746 non-null   float64
- 39  V162                            67246 non-null   float32
-dtypes: float32(18), float64(22)
-memory usage: 108.1 MB
