--- data-modelling/model_v5/list-addr-to-addr_target_encoded.json
@@ -1,57 +1,0 @@
-{
-    "32.0_other": 0.07407407407407407,
-    "60.0_other": 0.0962907455976021,
-    "65.0_other": 0.581081081081081,
-    "87.0_110.0": 0.01616915422885572,
-    "87.0_122.0": 0.08927173061863743,
-    "87.0_123.0": 0.025056565825861056,
-    "87.0_126.0": 0.02776807544533706,
-    "87.0_143.0": 0.012593094109681787,
-    "87.0_158.0": 0.025817555938037865,
-    "87.0_170.0": 0.04008016032064128,
-    "87.0_177.0": 0.02760143527463428,
-    "87.0_181.0": 0.02064528519300941,
-    "87.0_184.0": 0.013273957352059605,
-    "87.0_191.0": 0.017813765182186234,
-    "87.0_194.0": 0.011935208866155157,
-    "87.0_203.0": 0.050072046109510084,
-    "87.0_204.0": 0.027543186180422265,
-    "87.0_205.0": 0.01275264677574591,
-    "87.0_220.0": 0.03223270440251572,
-    "87.0_225.0": 0.02557100297914598,
-    "87.0_226.0": 0.047080979284369114,
-    "87.0_231.0": 0.031265930331350895,
-    "87.0_251.0": 0.026977950713359275,
-    "87.0_264.0": 0.017880593999393877,
-    "87.0_269.0": 0.03637919965760753,
-    "87.0_272.0": 0.02839743165420004,
-    "87.0_299.0": 0.022036365712328376,
-    "87.0_308.0": 0.07782101167315175,
-    "87.0_310.0": 0.020267009811806336,
-    "87.0_315.0": 0.018030030726418924,
-    "87.0_324.0": 0.0644966847498493,
-    "87.0_325.0": 0.024638137193203274,
-    "87.0_327.0": 0.018990346573824972,
-    "87.0_330.0": 0.0307973940666559,
-    "87.0_337.0": 0.023832814584259673,
-    "87.0_343.0": 0.016886930983847283,
-    "87.0_384.0": 0.03173374613003096,
-    "87.0_387.0": 0.020118733509234828,
-    "87.0_420.0": 0.03332085361287907,
-    "87.0_428.0": 0.02459016393442623,
-    "87.0_433.0": 0.022363092093678465,
-    "87.0_436.0": 0.00749375520399667,
-    "87.0_441.0": 0.023510466988727857,
-    "87.0_444.0": 0.008603573792190603,
-    "87.0_469.0": 0.03018018018018018,
-    "87.0_472.0": 0.025105254950881024,
-    "87.0_476.0": 0.03663973251602118,
-    "87.0_485.0": 0.017438364401683705,
-    "87.0_494.0": 0.006097560975609756,
-    "87.0_498.0": 0.023004828173814256,
-    "87.0_512.0": 0.029777486910994765,
-    "87.0_other": 0.019813519813519812,
-    "96.0_other": 0.10902255639097744,
-    "other": 0.0,
-    "other_other": 0.05675675675675676
-}
