--- data-modelling/model_v5/list-addr1-to-addr1_target_encoded.json
@@ -1,51 +1,0 @@
-{
-    "110.0": 0.01616915422885572,
-    "122.0": 0.08892355694227769,
-    "123.0": 0.02505236698785086,
-    "126.0": 0.027755957056821157,
-    "143.0": 0.012591389114541024,
-    "158.0": 0.02577319587628866,
-    "170.0": 0.04008016032064128,
-    "177.0": 0.02760143527463428,
-    "181.0": 0.02064528519300941,
-    "184.0": 0.013273957352059605,
-    "191.0": 0.017813765182186234,
-    "194.0": 0.011935208866155157,
-    "203.0": 0.050072046109510084,
-    "204.0": 0.027542305108601773,
-    "205.0": 0.01275264677574591,
-    "220.0": 0.03223270440251572,
-    "225.0": 0.02557100297914598,
-    "226.0": 0.04706831629908553,
-    "231.0": 0.031265930331350895,
-    "251.0": 0.026970954356846474,
-    "264.0": 0.017880593999393877,
-    "269.0": 0.03637919965760753,
-    "272.0": 0.028395552025416996,
-    "299.0": 0.02203070600993094,
-    "308.0": 0.07782101167315175,
-    "310.0": 0.02026375040205854,
-    "315.0": 0.018028985507246378,
-    "324.0": 0.06445783132530121,
-    "325.0": 0.024637361945816682,
-    "327.0": 0.018990346573824972,
-    "330.0": 0.03079242032730405,
-    "337.0": 0.023830695358349637,
-    "343.0": 0.016874541452677916,
-    "384.0": 0.03173374613003096,
-    "387.0": 0.020118733509234828,
-    "420.0": 0.03332085361287907,
-    "428.0": 0.02459016393442623,
-    "433.0": 0.022359154929577466,
-    "436.0": 0.00749375520399667,
-    "441.0": 0.0235074386552457,
-    "444.0": 0.008603573792190603,
-    "469.0": 0.030153015301530153,
-    "472.0": 0.025105254950881024,
-    "476.0": 0.03662952646239554,
-    "485.0": 0.017438364401683705,
-    "494.0": 0.006097560975609756,
-    "498.0": 0.023004828173814256,
-    "512.0": 0.029762878168438266,
-    "other": 0.06328583403185248
-}
