--- data-modelling/model_v5/list-addr2-to-addr2_target_encoded.json
@@ -1,8 +1,0 @@
-{
-    "32.0": 0.07228915662650602,
-    "60.0": 0.096002988419873,
-    "65.0": 0.581081081081081,
-    "87.0": 0.024207458667366913,
-    "96.0": 0.1070110701107011,
-    "other": 0.05289672544080604
-}
