--- data-modelling/model_v5/list-card2-to-card2_target_encoded.json
@@ -1,189 +1,0 @@
-{
-    "100.0": 0.01372514516980468,
-    "102.0": 0.0,
-    "103.0": 0.09163346613545817,
-    "104.0": 0.24180967238689546,
-    "105.0": 0.0,
-    "106.0": 0.0,
-    "110.0": 0.004484304932735426,
-    "111.0": 0.021859386152748038,
-    "113.0": 0.00423728813559322,
-    "122.0": 0.013043478260869565,
-    "127.0": 0.01296111665004985,
-    "130.0": 0.23429541595925296,
-    "133.0": 0.0,
-    "134.0": 0.02570694087403599,
-    "141.0": 0.25,
-    "142.0": 0.2672281776416539,
-    "143.0": 0.002066115702479339,
-    "148.0": 0.010309278350515464,
-    "152.0": 0.0,
-    "155.0": 0.02962962962962963,
-    "158.0": 0.060810810810810814,
-    "160.0": 0.09803921568627451,
-    "161.0": 0.023938223938223938,
-    "167.0": 0.0,
-    "170.0": 0.032153179190751446,
-    "174.0": 0.02722301474084641,
-    "176.0": 0.33585858585858586,
-    "177.0": 0.24382716049382716,
-    "181.0": 0.06407942238267147,
-    "184.0": 0.00398406374501992,
-    "191.0": 0.0,
-    "194.0": 0.008197965689995446,
-    "199.0": 0.007220216606498195,
-    "202.0": 0.1096938775510204,
-    "203.0": 0.009541984732824428,
-    "204.0": 0.08908582089552239,
-    "205.0": 0.03133903133903134,
-    "206.0": 0.004878048780487805,
-    "210.0": 0.012987012987012988,
-    "215.0": 0.017251078192387023,
-    "216.0": 0.057971014492753624,
-    "225.0": 0.03419746276889134,
-    "226.0": 0.016736401673640166,
-    "231.0": 0.06944444444444445,
-    "236.0": 0.004008016032064128,
-    "239.0": 0.05394190871369295,
-    "242.0": 0.012681159420289856,
-    "247.0": 0.03608247422680412,
-    "250.0": 0.021945137157107233,
-    "251.0": 0.0,
-    "253.0": 0.013215859030837005,
-    "254.0": 0.018333333333333333,
-    "258.0": 0.0,
-    "264.0": 0.02445777572681126,
-    "266.0": 0.16869300911854104,
-    "268.0": 0.005714285714285714,
-    "269.0": 0.02595797280593325,
-    "270.0": 0.0,
-    "271.0": 0.10951008645533142,
-    "272.0": 0.02912621359223301,
-    "275.0": 0.0,
-    "276.0": 0.012893982808022923,
-    "277.0": 0.07142857142857142,
-    "296.0": 0.19197138314785372,
-    "297.0": 0.003067484662576687,
-    "298.0": 0.017907634307257305,
-    "300.0": 0.007619047619047619,
-    "302.0": 0.0013422818791946308,
-    "303.0": 0.006289308176100629,
-    "304.0": 0.0,
-    "308.0": 0.0891089108910891,
-    "310.0": 0.03066989507667474,
-    "311.0": 0.25833333333333336,
-    "314.0": 0.0039603960396039604,
-    "318.0": 0.002386634844868735,
-    "320.0": 0.017921146953405017,
-    "321.0": 0.02838017260455853,
-    "322.0": 0.014084507042253521,
-    "324.0": 0.12315270935960591,
-    "325.0": 0.0048543689320388345,
-    "327.0": 0.07523245984784446,
-    "332.0": 0.008908685968819599,
-    "333.0": 0.003968253968253968,
-    "341.0": 0.20087336244541484,
-    "343.0": 0.025669642857142856,
-    "345.0": 0.024,
-    "346.0": 0.026134800550206328,
-    "348.0": 0.024390243902439025,
-    "350.0": 0.0,
-    "355.0": 0.009433962264150943,
-    "360.0": 0.010314778587942379,
-    "361.0": 0.015143603133159269,
-    "369.0": 0.004335260115606936,
-    "371.0": 0.009009009009009009,
-    "372.0": 0.0,
-    "375.0": 0.07385952208544533,
-    "381.0": 0.022222222222222223,
-    "383.0": 0.022222222222222223,
-    "385.0": 0.040968342644320296,
-    "387.0": 0.036585365853658534,
-    "390.0": 0.017346460384435068,
-    "391.0": 0.0,
-    "393.0": 0.026592022393282014,
-    "396.0": 0.005333333333333333,
-    "399.0": 0.028164347249834326,
-    "400.0": 0.022727272727272728,
-    "404.0": 0.07285307285307285,
-    "408.0": 0.07510008006405124,
-    "411.0": 0.01910828025477707,
-    "414.0": 0.026578073089700997,
-    "417.0": 0.053854875283446714,
-    "418.0": 0.013828238719068414,
-    "420.0": 0.006611570247933884,
-    "422.0": 0.043668122270742356,
-    "428.0": 0.003110419906687403,
-    "431.0": 0.002403846153846154,
-    "432.0": 0.0,
-    "440.0": 0.18461538461538463,
-    "442.0": 0.021052631578947368,
-    "445.0": 0.033112582781456956,
-    "449.0": 0.024054982817869417,
-    "452.0": 0.012274959083469721,
-    "453.0": 0.05945121951219512,
-    "454.0": 0.008403361344537815,
-    "455.0": 0.01349527665317139,
-    "459.0": 0.18893129770992367,
-    "460.0": 0.01195219123505976,
-    "462.0": 0.01824817518248175,
-    "464.0": 0.0,
-    "470.0": 0.010158013544018058,
-    "472.0": 0.0,
-    "474.0": 0.04938271604938271,
-    "476.0": 0.006471716203259827,
-    "477.0": 0.009569377990430622,
-    "478.0": 0.037848605577689244,
-    "479.0": 0.024553571428571428,
-    "480.0": 0.03211009174311927,
-    "481.0": 0.03651226158038147,
-    "489.0": 0.014522821576763486,
-    "490.0": 0.02374679578467673,
-    "492.0": 0.008097165991902834,
-    "494.0": 0.020043572984749455,
-    "496.0": 0.002967359050445104,
-    "498.0": 0.013043478260869565,
-    "500.0": 0.18477990549614523,
-    "507.0": 0.04526748971193416,
-    "512.0": 0.018668152688771245,
-    "513.0": 0.008985879332477536,
-    "514.0": 0.04757194244604317,
-    "517.0": 0.0035671819262782403,
-    "520.0": 0.012953367875647668,
-    "523.0": 0.00851063829787234,
-    "528.0": 0.08789954337899543,
-    "529.0": 0.16188870151770657,
-    "532.0": 0.023178807947019868,
-    "533.0": 0.004975124378109453,
-    "536.0": 0.11004784688995216,
-    "543.0": 0.005988023952095809,
-    "545.0": 0.08129123468426014,
-    "548.0": 0.11152416356877323,
-    "549.0": 0.060240963855421686,
-    "550.0": 0.03169014084507042,
-    "551.0": 0.0,
-    "553.0": 0.049811470529867036,
-    "554.0": 0.012379642365887207,
-    "555.0": 0.02275162286506586,
-    "556.0": 0.0029940119760479044,
-    "558.0": 0.0009505703422053232,
-    "559.0": 0.0,
-    "561.0": 0.00964630225080386,
-    "562.0": 0.03356831151393085,
-    "564.0": 0.0,
-    "565.0": 0.0653061224489796,
-    "566.0": 0.04326923076923077,
-    "567.0": 0.015022860875244938,
-    "568.0": 0.0,
-    "571.0": 0.0,
-    "572.0": 0.02666666666666667,
-    "579.0": 0.05680119581464873,
-    "583.0": 0.04028477166174683,
-    "584.0": 0.006802721088435374,
-    "585.0": 0.09055118110236221,
-    "586.0": 0.009009009009009009,
-    "593.0": 0.0,
-    "595.0": 0.021015761821366025,
-    "599.0": 0.006968641114982578,
-    "other": 0.0227793596777554
-}
