--- data-modelling/model_v5/list-card3-to-card3_target_encoded.json
@@ -1,15 +1,0 @@
-{
-    "102.0": 0.02601156069364162,
-    "106.0": 0.012605042016806723,
-    "117.0": 0.018207282913165267,
-    "119.0": 0.09545454545454546,
-    "143.0": 0.03767660910518053,
-    "144.0": 0.06936416184971098,
-    "146.0": 0.009355509355509356,
-    "147.0": 0.02564102564102564,
-    "150.0": 0.024859537578191603,
-    "162.0": 0.012396694214876033,
-    "185.0": 0.12597313908490781,
-    "203.0": 0.0049504950495049506,
-    "other": 0.04390243902439024
-}
