--- data-modelling/model_v5/list-card5-to-card5_target_encoded.json
@@ -1,33 +1,0 @@
-{
-    "100.0": 0.018617021276595744,
-    "102.0": 0.06610582668132499,
-    "117.0": 0.015205913410770855,
-    "118.0": 0.02009273570324575,
-    "126.0": 0.011976047904191617,
-    "129.0": 0.036,
-    "137.0": 0.14420028424620093,
-    "138.0": 0.08355074354291678,
-    "141.0": 0.07264957264957266,
-    "142.0": 0.043689320388349516,
-    "146.0": 0.03770197486535009,
-    "147.0": 0.08814589665653495,
-    "150.0": 0.018699910952804988,
-    "162.0": 0.07048331415420023,
-    "166.0": 0.010512085401953244,
-    "183.0": 0.014925373134328358,
-    "185.0": 0.035,
-    "190.0": 0.022452504317789293,
-    "194.0": 0.035398230088495575,
-    "195.0": 0.03890558279297802,
-    "197.0": 0.05302402651201325,
-    "198.0": 0.04861111111111111,
-    "202.0": 0.008165563846262143,
-    "203.0": 0.0877742946708464,
-    "219.0": 0.06356529539920544,
-    "223.0": 0.05548387096774193,
-    "224.0": 0.037850076290790964,
-    "226.0": 0.029549689021700975,
-    "229.0": 0.02054794520547945,
-    "236.0": 0.034545454545454546,
-    "other": 0.03212290502793296
-}
