--- data-modelling/model_v5/model01_xgboost_joblib_describe_X.csv
@@ -1,9 +1,0 @@
-,V264,addr2_target_encoded,V283,C1,C5,R_emaildomain_bin_google,M4_nan,V30,V156,D3,card2_target_encoded,V67,V198,M5_T,device_name_Trident,ProductCD_R,V62,id_12_Found,M6_F,V44,D2,V54,P_emaildomain_bin_anonymous,M9_F,addr_target_encoded,D8,card3_target_encoded,Browser_id_31_Internet Explorer,V13,DeviceType_mobile,V307,M3_F,V86,TransactionAmt,V289,OS_id_30_Mac,V312,V160,D4,card6_debit,V142,V47,V80,TransactionAmt_decimal,P_emaildomain_suffix_us,D15
-count,25169.0,132618.0,146640.0,147635.0,147635.0,147635.0,147635.0,135154.0,14699.0,90893.0,145464.0,135138.0,28383.0,147635.0,147635.0,147635.0,135138.0,147635.0,147635.0,108937.0,85630.0,135138.0,147635.0,147635.0,132595.0,16180.0,146877.0,147635.0,135154.0,147635.0,147635.0,147635.0,130806.0,147635.0,146640.0,147635.0,147635.0,14705.0,108953.0,147635.0,14699.0,108937.0,130806.0,147635.0,147635.0,130836.0
-mean,428.30450439453125,0.024535479084980136,0.9524208903312683,9.599166864225962,6.039218342533952,0.0826971924001761,0.45863108341517933,0.41296595335006714,0.7567861676216125,27.537313104419482,0.03288134623189912,0.9924595355987549,0.93925940990448,0.19328749957665864,0.008256849663020286,0.044935144105395064,0.8614823222160339,0.04044433907948657,0.41469163816168253,1.069948673248291,178.115870606096,0.6905385851860046,0.058962983032478745,0.09798489518068208,0.024503740732063375,131.90248226590538,0.0335463937110827,0.02322619974938192,0.6606242060661316,0.08059064584956142,922.394287109375,0.1526399566498459,1.0608917474746704,137.11464901954145,0.24918848276138306,0.015043858163714566,47.607791900634766,2296.007080078125,159.61609134213836,0.7658888474955126,0.06435812264680862,1.0328079462051392,0.1263168305158615,398.94019710773193,0.040349510617401024,188.73421688220367
-std,3135.8291015625,0.006312233610127523,1.5482850074768066,57.84887306545564,27.465982162494356,0.2754249091794666,0.4982873613943747,0.5582315921783447,0.8785057663917542,62.55484999797557,0.03629633115652058,0.2647111117839813,0.5346589684486389,0.39487782700444657,0.09049159940407524,0.2071624184578666,0.5112821459770203,0.1969996380357448,0.4926704045130112,0.3952120244503021,189.86755908285207,0.5280020236968994,0.23555620454913434,0.2972952306476126,0.010366506938500273,214.51008507658426,0.028249682729696222,0.15062170183412918,0.5255631804466248,0.272206347353479,8334.29296875,0.3596413162805534,0.35698962211608887,245.22100092455324,0.5991319417953491,0.12172773250983383,173.45123291015625,9678.95703125,211.33813409544572,0.4234434263139928,0.39567649364471436,0.20425327122211456,0.3792712688446045,441.1711980665748,0.19677827596108255,224.23245820374515
-min,0.0,0.024207458667366913,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0049504950495049506,0.0,0.0,0.0,0.0,0.0,0.0,0.364,0.0,0.0,0.0,0.0,-74.0,0.0,0.0,0.0,0.0,0.0,0.0,-74.0
-25%,0.0,0.024207458667366913,0.0,1.0,0.0,0.0,0.0,0.0,0.0,1.0,0.018668152688771245,1.0,1.0,0.0,0.0,0.0,1.0,0.0,0.0,1.0,29.0,0.0,0.0,0.0,0.018030030726418924,1.7083330154418943,0.024859537578191603,0.0,0.0,0.0,0.0,0.0,1.0,43.504999999999995,0.0,0.0,0.0,0.0,0.0,1.0,0.0,1.0,0.0,0.0,0.0,0.0
-50%,0.0,0.024207458667366913,1.0,1.0,0.0,0.0,0.0,0.0,1.0,8.0,0.0227793596777554,1.0,1.0,0.0,0.0,0.0,1.0,0.0,0.0,1.0,98.0,1.0,0.0,0.0,0.023832814584259673,27.70833206176757,0.024859537578191603,0.0,1.0,0.0,21.076799392700195,0.0,1.0,67.95,0.0,0.0,0.0,0.0,42.0,1.0,0.0,1.0,0.0,0.0,0.0,75.0
-75%,50.0,0.024207458667366913,1.0,3.0,1.0,0.0,1.0,1.0,1.0,26.0,0.032153179190751446,1.0,1.0,0.0,0.0,0.0,1.0,0.0,1.0,1.0,280.0,1.0,0.0,0.0,0.027543186180422265,169.9166717529297,0.024859537578191603,0.0,1.0,0.0,187.89999389648438,0.0,1.0,117.0,0.0,0.0,0.0,30.0,285.0,1.0,0.0,1.0,0.0,949.0,0.0,357.0
-max,104060.0,0.581081081081081,56.0,1505.0,335.0,1.0,1.0,6.0,24.0,819.0,0.33585858585858586,7.0,21.0,1.0,1.0,1.0,10.0,1.0,1.0,32.0,639.0,4.0,1.0,1.0,0.581081081081081,1638.875,0.12597313908490781,1.0,6.0,1.0,145765.0,1.0,13.0,5366.82,9.0,1.0,12500.0,161644.296875,869.0,1.0,5.0,6.0,15.0,998.0,1.0,879.0
