--- data-modelling/model_v5/model01_xgboost_joblib_describe_y.csv
@@ -1,9 +1,0 @@
-,0
-count,147635.0
-mean,0.0313949944118942
-std,0.17438335561549956
-min,0.0
-25%,0.0
-50%,0.0
-75%,0.0
-max,1.0
