--- data-modelling/model_v5/model02_lightgbm_joblib_describe_X.csv
@@ -1,9 +1,0 @@
-,card1,TransactionAmt,card2_target_encoded,addr_target_encoded,addr1_target_encoded,dist1,D15,id_02,TransactionAmt_decimal,C1,D4,card5_target_encoded,D2,D10,D11,V307,D1,D8,D5,V310,D3,id_05,V127,id_06,V314,D9,V264,V312,id_01,D14,V203,C5,D6,V283,D12,V36,V96,V221,V62,V82,V282,V54,V76,V37,V20,V5,V44,V285,V77,V56
-count,442905.0,442905.0,436143.0,392207.0,392207.0,171642.0,370591.0,112436.0,442905.0,442905.0,312665.0,439850.0,224113.0,379342.0,208084.0,442893.0,442631.0,58746.0,204555.0,442893.0,236769.0,109424.0,442904.0,109424.0,442631.0,58746.0,105261.0,442893.0,115052.0,47457.0,111362.0,442905.0,55607.0,442631.0,50118.0,312634.0,442904.0,112961.0,378306.0,370570.0,442631.0,378306.0,370570.0,312634.0,379313.0,208084.0,312634.0,442893.0,370570.0,378306.0
-mean,9862.64624016437,134.2162646278547,0.034800512675888406,0.024956209348634774,0.024956209348634774,123.00763799070158,154.9220650258641,169408.0950585222,372.7202899041555,15.590221379302559,133.167757823869,0.03502557690121632,166.29540455038307,116.79826119965625,134.9482372503412,237.4404296875,89.20036102306436,149.95689256268827,42.31998728948204,109.3085708618164,28.65277549003459,1.733568504167276,156.15206909179688,-6.588810498610908,40.471561431884766,0.5602526994679473,147.46400451660156,36.36254119873047,-9.891527309390536,50.50959816254715,452.88275146484375,5.415628633679908,62.6738180444908,1.003933310508728,51.32186040943374,0.5519905090332031,1.2680941820144653,1.2907906770706177,0.869735062122345,0.8446636199951172,0.8315210342407227,0.5947777628898621,0.5640580654144287,1.113647222518921,0.8511229753494263,0.8716143369674683,1.0887491703033447,1.0724396705627441,1.0897374153137207,1.1137306690216064
-std,4896.986456472178,227.66548850867832,0.03921165631781867,0.013336997213738084,0.010012538160691212,377.3848052682449,193.79939603007523,151403.46184110566,431.3330835126008,150.54276684015534,183.0267806342511,0.0228223473414256,172.16707744781837,174.60169490202384,175.46205011269467,1564.725341796875,151.7199946705314,236.02230502454512,88.0502234091896,333.30859375,62.31665051157136,5.278705424917757,763.694091796875,16.462445673660785,163.98477172851562,0.31883208493959264,2025.41015625,171.59320068359375,14.223173689948403,127.54221878109205,4866.63134765625,25.200565939973874,136.90513372627328,1.5619667768478394,119.04720335936466,0.5396079421043396,4.4840803146362305,2.856454372406006,0.4735829532146454,0.42279526591300964,0.9164148569107056,0.5347328782081604,0.5391847491264343,0.7403247356414795,0.4629403054714203,0.47199803590774536,0.7045078873634338,3.160334587097168,0.5558487176895142,0.5234606266021729
-min,1000.0,0.251,0.0,0.0,0.006097560975609756,0.0,-83.0,30.0,0.0,0.0,-122.0,0.008165563846262143,0.0,0.0,-53.0,0.0,0.0,0.0,0.0,0.0,0.0,-72.0,0.0,-100.0,0.0,0.0,0.0,0.0,-100.0,-193.0,0.0,0.0,-83.0,0.0,-83.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0
-25%,6019.0,43.25,0.018668152688771245,0.018030030726418924,0.018028985507246378,3.0,0.0,69379.5,0.0,1.0,0.0,0.029549689021700975,24.0,0.0,0.0,0.0,0.0,0.9583330154418944,1.0,0.0,1.0,0.0,0.0,-5.0,0.0,0.2083330005407333,0.0,0.0,-10.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,1.0,1.0,1.0,0.0,0.0,0.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0
-50%,9633.0,68.95,0.02374679578467673,0.023832814584259673,0.023830695358349637,9.0,43.0,123370.0,0.0,1.0,21.0,0.029549689021700975,97.0,11.0,32.0,0.0,1.0,41.20833206176758,9.0,0.0,8.0,0.0,0.0,0.0,0.0,0.6666659712791443,0.0,0.0,-5.0,0.0,0.0,0.0,0.0,1.0,0.0,1.0,0.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,0.0,1.0,1.0
-75%,14128.0,125.0,0.03356831151393085,0.02760143527463428,0.027755957056821157,25.0,300.0,218645.75,949.0,3.0,245.0,0.037850076290790964,275.0,184.0,252.0,139.9499969482422,113.0,192.23958206176758,32.0,98.0,27.0,1.0,99.8499984741211,0.0,0.0,0.8333330154418945,29.732999801635742,0.0,-5.0,0.0,24.53059959411621,1.0,21.0,1.0,10.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0,1.0
-max,18396.0,6450.97,0.33585858585858586,0.581081081081081,0.08892355694227769,5431.0,821.0,999595.0,999.0,4685.0,815.0,0.14420028424620093,640.0,818.0,620.0,108800.0,640.0,1707.7916259765625,775.0,55125.0,729.0,52.0,160000.0,0.0,6159.89013671875,0.9583330154418944,153600.0,55125.0,0.0,815.0,101640.0,349.0,821.0,68.0,610.0,4.0,126.0,384.0,7.0,7.0,32.0,6.0,6.0,54.0,15.0,6.0,48.0,95.0,30.0,29.0
