--- data-modelling/model_v5/model02_lightgbm_joblib_describe_y.csv
@@ -1,9 +1,0 @@
-,0
-count,147635.0
-mean,0.024140617062349715
-std,0.15348617931089295
-min,0.0
-25%,0.0
-50%,0.0
-75%,0.0
-max,1.0
